requirements.txt
Django==3.1.5
django-ninja==0.10.1
